B 1086 add infinite scroll in the project list (#96)
* WIP

* Refactoring get projects for anonymous user

* Fixing ITs for get projects given an anonymous user

* Refactoring get projects for an authenticated user

* Fixing build

+++ domain/src/main/java/onlydust/com/marketplace/api/domain/model/UserProfile.java
@@ -17,7 +17,7 @@ public class UserProfile {
     private String bio;
     private String website;
     @Builder.Default
-    private Map<String, Integer> technologies = new HashMap<>();
+    private Map<String, Long> technologies = new HashMap<>();
     private UserProfileCover cover;
     @Builder.Default
     private List<Contact> contacts = new ArrayList<>();
